Greatly improve altitude generation
Rewrite altitude generation so that it is MUCH faster and now occurs by
default, no longer depend on geocoder for altitude generation,
randomize altitudes before visits so that they don't stay the same for
each point, provide some more documentation of config examples.

Note: ensure that all of the dependencies are installed and up to date,
until recently pgoapi was putting altitude in the wrong field.