Optionally complete tutorials automatically
If COMPLETE_TUTORIAL is set: run through all the tutorial steps,
configure the avatar, catch the starter pokemon, and set it as a buddy
for all accounts that have not yet done that. Also add parameters to
call_chain to allow it to be used more in the app simulation process.